:whale: [#18] Fix docker setup with Open Zaak hostname
Open Zaak and WooP can't share the host.docker.internal hostname in
ALLOWED_HOSTS, as that makes Open Zaak treat the
host.docker.internal:8000 URLs as internal (while they are external
and provided by the WooP backend).

The simplest workaround for this is to add an additional network alias
pointing to the same IP address, and expect developers to set up the
host name in their /etc/hosts file.