:art: [#11] Simplify implementation
The custom list-masquerading-as-a-queryset is not type-safe or sound,
and the necessity to override the get_object implementation to be able
to do UUID lookups is a manifestation of that.

I've decided that for now we take the performance hit while we still
have small data sets for the themes - looking up the children of each
node involves an additional query. See #63 for the ticket to optimize
this.

Some initial experiments show that it's not a simple or obvious
solution to optimize this with Django and it may involve custom DB
operations, querysets and managers with a custom model iterable. The
cost then will be memory usage to be able to properly resolve the
parent-child relations.