:white_check_mark: [#45] Update test to hit real endpoint via LiveServerTestCase
It's too confusing to have to include a fixture in docker for a
'regular' django test, instead it makes more sense to let Open Zaak
talk to the real server thread from LiveServerTestCase.

However, this requires us to expose the application on an additional
port, otherwise we get multiple programs trying to bind the same
port number on the host machine.