:zap: [#16] Avoid GFK's in admin list triggering too many queries
log.content_object cannot be prefetch in Django 4.2 (though it looks
like 5.2 will have it). To avoid doing a query for each log record on a
page in the admin list (100 records -> 100 queries), we can cache the
representation and make sure we try the cache first before doing extra
queries.