:art: [#16] Implement permission class tests as unit test
We don't need to test the entire views, especially in a general-purpose
piece of functionality you want to provide unit tests and separately
implement the integration tests in the places where it should be
used.

This cleanup removes the tight coupling with the publications or
metadata apps and instead implements the expected behaviour of
the permission class itself.