:truck: [#16] Move tests into right place
Tests that operate on a particular model or django app belong in
that particular django app - not in the general purpose app
providing the implementation.

Ideally, the general purpose implementation has agnostic
unit tests to ensure the correct behaviour too, but with
admin mixins that's a bit harder to achieve, so we'll
pick the lesser evil here.