:sparkles: [#16] Implement admin and API logging hooks (GH PR #74)
* :sparkles: [#16] added logger function

* :sparkles: [#16] created generic auditlog mixin for admin

* :sparkles: [#16] added audit log event list filter

* :sparkles: [#16] added audit log request headers

* :sparkles: [#16] extended auditlog to create logs for inline admin

* :sparkles: [#16] added system check for missing AdminAuditLogMixin

* :sparkles: [#16] added mixin for api loggin

* :white_check_mark: [#16] added unit tests for audit logging and api header requirement

* :white_check_mark: [#16] added inline admin audit log testing

* :art: [#16] improved type hints

* :art: [#16] Type annotate + solidify audit headers hook impl

The paths key in OpenAPI schema can have keys/values that are not
operations, so blindly adding the parameters to each item is too
fragile.

Additionally, the permission class can re-use the headers defined
for the API spec generation, so that we guarantee that
implementation and documentation actually remain properly in sync.

* :rewind: [#16] Revert audit headers requirement for Catalogi API endpoint

This endpoint is implemented using the ZGW API's Catalogi API spec,
and the read endpoint does not require any kind of audit headers.

* :label: [#16] Fix type checker annotations

Created an upstream issue to django-stubs for the incorrect method
signature compared to Django's actual implementation:
https://github.com/typeddjango/django-stubs/issues/2413

* :label: [#16] Clean up and simplify types used in logging module

* Declared the JSON types in general purpose module, they're not
  logging specific and will get used in a broader context.
* Used the new type alias syntax which greatly declutters the code
* Removed TYPE_CHECKING guards as the circular imports/AppRegistry not
  ready errors are resolved earlier
* Made the logevent functions kwarg-only, since positional arguments
  are too confusing at call-sites.

* :art: [#16] Clean up logging implementation

* There was no request to log status codes, so don't do it
* When snapshotting the model data, don't dump to JSON and
  then load it again, instead use a variant of django's
  built in django.forms.model_to_dict and rely on Django
  to serialize it at the DB level
* Removed most of the # type: ignore comments and replaced
  with proper type annotations
* Made the top-level API for the logging module more explicit,
  relying on positional arguments is too fragile and makes
  it unreadable what is semantically being logged.
* Moved the public API into woo_publications.logging.service,
  since doing model imports in the __init__.py layer tries
  to use django concepts before the app-registry is ready,
  and that prevents us from having properly typed code.

* :art: [#16] Enable audit trail logging on user model

* :art: [#16] Clean up tests

There's no need for the audit headers to be an instance variable,
instead they can be declared as module constants. The indirection
to setUp is there already anyway, and the constant notation makes it
more clear that nothing special is going on.

* :truck: [#16] Put API schema generation tests in correct package

* :art: [#16] Rewrite test to be more semantic

The building blocks are already there - all we care about is that
the correct amount of log records are shown, and that the correct
records are shown.

* :art: [#16] Apply consistent kwargs-only Python API

* :art: [#16] Clean up admin logging tests

* :art: [#16] Properly add dynamic webtest fields

This also resolves the type checker errors.

* :white_check_mark: [#16] Add test for incomplete user information

* :art: [#16] Implement permission class tests as unit test

We don't need to test the entire views, especially in a general-purpose
piece of functionality you want to provide unit tests and separately
implement the integration tests in the places where it should be
used.

This cleanup removes the tight coupling with the publications or
metadata apps and instead implements the expected behaviour of
the permission class itself.

* :truck: [#16] Move tests into right place

Tests that operate on a particular model or django app belong in
that particular django app - not in the general purpose app
providing the implementation.

Ideally, the general purpose implementation has agnostic
unit tests to ensure the correct behaviour too, but with
admin mixins that's a bit harder to achieve, so we'll
pick the lesser evil here.

* :truck: [#16] Move publication API logging tests to correct location

* :white_check_mark: [#16] Add missing tests for metadata audit logs

* :art: [#16] Implement audit headers as DRF spectacular extension

Instead of a post processing hook. The extension approach allows us to add it to the
API schema only when it's relevant - e.g. the Catalogi API endpoint does not
require this.

* :bug: [#16] Fix inline update logging

This was incorrectly being logged as create

* :pencil: [#16] Ensure Python API is documented properly

---------

Co-authored-by: Sergei Maertens <sergei@maykinmedia.nl>