:art: Clean up tests with side-effects
Calling the real management command implementation leads to changed
UUIDs in fixture files produced by tests, which in turn leads to
cluttered git diff output and a risk for updated assets that shouldn't
be updated. The UUIDs are stored in remote systems once they are loaded
in an environment and should ideally not change.

The VCR cassettes can be removed since they don't reflect the actual
test any longer - we care about the interface between management
command and actual implementation, and the latter is tested separately.