:sparkles: [#26] Ensure a default is set for the is_actief flag in filters
@bart-maykin was right

A default can be set in the filter class, but to only apply this in the
list action we need access to the viewset instance (to check
view.action), which is not passed along in the kwargs by django-filters.

Adding that requires us to use a custom filter backend, and that in turn
affects all the filter sets again that need to handle this kwarg, which
becomes way too much effort/work for the gains it gives.

Co-authored-by: bart-maykin <bart.burgman@maykinmedia.nl>