:card_file_box: [#19] Add global configuration settings
While there's technically no reason for us to limit the documents API
service configuration to a single service, this requires more upfront
thought and work in the integration with the publication component or
some kind of routing logic/setup, and that's out of scope of the
project for the time being. So, this global configuration allows you
to set up one service and one RSIN for the documents API that will be
used to store the binary data (and some metadata) of our Woo publication
documents.