:art: [#19] Ensure DRF router uses django path instead of re_path
Using the new paths + path converters makes it easier to deal with
viewset actions for additional routes without needing to validate
the path parameter manually, and keeps everything consistent.

This requires disabling the 'format' endpoint variations, otherwise
they show up in the API specification (looks like drf-spectacular can't
filter those out properly yet), which requires setting the attribute
on the router as there is no __init__ kwarg for it.

For proper type casting and annotations, we need to specify the
converter on the viewset too.