:sparkles: [#19] Implement viewset endpoint for file part upload
Wire up the detail route to accept PUT requests with file part upload
data. The parser must be set to multipart to properly handle the file
uploads from the request data, and accordingly the serializer must
mark the appropriate fields as read/write only.

We opt to return a generic status code of 204, as the response data
of the upstream request has little value and this simplifies having to
relay the upstream 200 or 201 depending on whether the part data was
created or updated.