:bug: Avoid crashing in list/detail document endpoints
In theory we could look up the zgw document from the documenten API
reference, and return back the file parts from the upstream API,
however this isn't really needed since they are already returned on the
initial create call.

Having this information available for the sitemap generation also adds
no benefits at all, and doing the lookups is a performance problem,
so if we can get away with not looking up the information at all, that's
the preferred approach.