:wrench: Ignore coverage on code sourced from default-project
This is proven code bootstrapped from default-project (let's hope these
aren't famous last words) that don't ship with tests out of the box.

There are talks of putting these utilities into a separate utility
package and have them tested there, and make default-project pull in
this dependency so that we don't have to repeat this coverage workaround
for every project that we start, but that's not helping us in the
short term.