:wrench: [#164] Tweak timeouts and disable nginx buffering
This should help address the write timeout errors, as explained in
https://kmiku7.github.io/2018/02/02/Solve-the-problem-of-uwsgi-uwsgi-response-write-body-do-TIMEOUT-error/

'uwsgi_response_write_body_do() TIMEOUT !!!' presumably happens when
uwsgi is writing to the nginx buffers, which nginx cannot empty quickly
enough due to the limited download speed of the client (the browser,
connected with nginx). This buffer size seems to sit around 1GB, which
is what we observe where the request fails/gets cut off.

We're exposing uwsgi as an http server rather than using the uwsgi
protocol or a socket directly, so in theory we should be able to
tweak these timeouts using --http-timeout and --harakiri rather than
the --socket-timeout which would not have any effect, but not
specifying it revealed errors as early as 30s into the request (
on 4G speeds the download time is nearly 20 minutes).

Tweaking the timeouts alone didn't seem to resolve the problem
fully, and it's also fragile because larger files will take even
more time to download. So, after some experimenting with the
X-Accel-Buffering header to disable temporary file buffering
in nginx entirely, that seems to be the most robust solution.

Additionally, added the recommended flags from Sentry SDK output for
threading support.